✨feat: 재사용성을 위해 사용자 프로필과 닉네임 분리
